requirements.txt
faiss-cpu
groq
langchain-groq
PyPDF2
langchain
streamlit
langchain_community
python-dotenv
pypdf
sentence-transformers
langchain-core 
langchain-community 
pytesseract 
pdf2image 
poppler-utils
pillow
